requirements.txt
pandas>=2.1.0
openpyxl>=3.1.0
python-docx>=1.1.0
PyPDF2>=3.0.0
streamlit>=1.30.0
tkinterdnd2==0.3.0

streamlit==1.30.0
